Setup TSC and ensure testing works
