+++ lib/semantic.js
