+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"compilerOptions": {
+		"target": "es5",
+		"lib": ["es6"],
+		"module": "es6",
+	},
+	"include": ["lib/*"]
+}
